Add static_cast for mixed signedness comparisons
Cast int values to size_t and R_xlen_t values to size_t in comparison operations to eliminate remaining signedness warnings on Windows builds.

+++ inst/include/interface/rcpp/rcpp_objects/rcpp_fleet.hpp
@@ -516,7 +516,7 @@ class FleetInterface : public FleetInterfaceBase {
           this->age_to_length_conversion.size());
 
       if (this->age_to_length_conversion.size() !=
-          (this->n_ages.get() * this->n_lengths.get())) {
+          static_cast<size_t>(this->n_ages.get() * this->n_lengths.get())) {
         FIMS_ERROR_LOG(
             "age_to_length_conversion don't match, " +
             fims::to_string(this->age_to_length_conversion.size()) + " != " +
